style: enforce file endings
