+++ Sources/Mailozaurr/DmarcReports/DmarcReport.cs
@@ -16,21 +16,3 @@ public sealed class DmarcReport {
     /// <summary>Collection of zipped XML attachments.</summary>
     public IList<DmarcReportAttachment> Attachments { get; } = new List<DmarcReportAttachment>();
 }
-
-/// <summary>
-/// Represents a zipped XML attachment belonging to a DMARC report.
-/// </summary>
-public sealed class DmarcReportAttachment : IDisposable {
-    /// <summary>File name of the attachment.</summary>
-    public string Name { get; }
-
-    /// <summary>Stream containing the zipped attachment.</summary>
-    public Stream Content { get; }
-
-    public DmarcReportAttachment(string name, Stream content) {
-        Name = name;
-        Content = content;
-    }
-
-    public void Dispose() => Content.Dispose();
-}
